+++ loadit/table_data.py
@@ -28,11 +28,11 @@ class TableData(object):
 
     @property
     def LIDs(self):
-        return np.array(self._LIDs)
+        return np.array(self._LIDs, dtype=self._LIDs.dtype)
 
     @property
     def IDs(self):
-        return np.array(self._IDs)
+        return np.array(self._IDs, dtype=self._IDs.dtype)
 
     @property
     def fields(self):
